# docs/api/consolidated_api_docs.html/auth-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Auth Service",
    "description": "Authentication and authorization service for Cryptobot",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/api-keys": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Api Keys",
        "description": "List API keys for current user",
        "operationId": "list_api_keys_auth_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyOut"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Auth Api Keys Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key for exchanges",
        "operationId": "create_api_key_auth_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth__models__user__APIKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth__schemas__token__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/auth/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Delete Api Key",
        "description": "Delete an API key",
        "operationId": "delete_api_key_auth_api_keys__key_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api-keys/api-keys": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "List Api Keys",
        "description": "List all API keys for the current user\n\nReturns a list of all API keys owned by the current user,\nincluding active, rotating, and expired keys.",
        "operationId": "list_api_keys_api_keys_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key\n\nCreates a new API key for the specified exchange with the given description.\nThe key will expire after the specified number of days (default: 90).",
        "operationId": "create_api_key_api_keys_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api-keys/api-keys/rotate": {
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Rotate Api Key",
        "description": "Rotate an API key\n\nCreates a new API key to replace the specified key.\nThe old key will remain valid for the specified grace period (default: 24 hours).",
        "operationId": "rotate_api_key_api_keys_api_keys_rotate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyRotate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api-keys/api-keys/revoke": {
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Revoke Api Key",
        "description": "Revoke an API key\n\nImmediately invalidates the specified API key.\nThis action cannot be undone.",
        "operationId": "revoke_api_key_api_keys_api_keys_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyRevoke"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api-keys/api-keys/emergency-revoke": {
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Emergency Revoke Api Key",
        "description": "Emergency revocation for compromised API key\n\nImmediately invalidates the specified API key and marks it as compromised.\nThis will trigger security alerts and create a high-severity audit log entry.",
        "operationId": "emergency_revoke_api_key_api_keys_api_keys_emergency_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCompromise"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api-keys/api-keys/history/{exchange}": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Get Api Key History",
        "description": "Get API key version history\n\nReturns the version history of API keys for the specified exchange.",
        "operationId": "get_api_key_history_api_keys_api_keys_history__exchange__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exchange",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exchange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/api-keys/expiring": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Get Expiring Api Keys",
        "description": "Get API keys that are expiring soon\n\nReturns a list of API keys that will expire within the specified number of days.",
        "operationId": "get_expiring_api_keys_api_keys_api_keys_expiring_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyExpiringResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/api-keys/{key_id}": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Get Api Key",
        "description": "Get API key details\n\nReturns detailed information about the specified API key.",
        "operationId": "get_api_key_api_keys_api_keys__key_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/api-keys/validate": {
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Validate Api Key",
        "description": "Validate an API key\n\nChecks if the provided API key is valid and returns its details.\nThis endpoint is primarily for internal service-to-service authentication.",
        "operationId": "validate_api_key_api_keys_api_keys_validate_post",
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIKeyListResponse": {
        "properties": {
          "keys": {
            "items": {
              "$ref": "#/components/schemas/auth__schemas__api_key__APIKeyResponse"
            },
            "type": "array",
            "title": "Keys",
            "description": "List of API keys"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of keys"
          },
          "active": {
            "type": "integer",
            "title": "Active",
            "description": "Number of active keys"
          },
          "rotating": {
            "type": "integer",
            "title": "Rotating",
            "description": "Number of keys in rotation"
          },
          "expiring_soon": {
            "type": "integer",
            "title": "Expiring Soon",
            "description": "Number of keys expiring within 7 days"
          }
        },
        "type": "object",
        "required": [
          "keys",
          "total",
          "active",
          "rotating",
          "expiring_soon"
        ],
        "title": "APIKeyListResponse",
        "description": "Response containing a list of API keys"
      },
      "APIKeyOut": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "is_test": {
            "type": "boolean",
            "title": "Is Test",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "description",
          "exchange",
          "id",
          "key"
        ],
        "title": "APIKeyOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "auth__models__user__APIKeyCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "is_test": {
            "type": "boolean",
            "title": "Is Test",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "description",
          "exchange"
        ],
        "title": "APIKeyCreate"
      },
      "auth__schemas__api_key__APIKeyCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of the key's purpose"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "Exchange this key is for"
          },
          "is_test": {
            "type": "boolean",
            "title": "Is Test",
            "description": "Whether this is a test key",
            "default": false
          },
          "expiry_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Days",
            "description": "Days until key expires",
            "default": 90
          }
        },
        "type": "object",
        "required": [
          "description",
          "exchange"
        ],
        "title": "APIKeyCreate",
        "description": "Request to create a new API key"
      },
...
      "auth__schemas__token__APIKeyResponse": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "created_at"
        ],
        "title": "APIKeyResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}